Backend/requirements_prod.txt
# Production Requirements
asgiref==3.9.2
certifi==2025.10.5
charset-normalizer==3.4.3
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
mongoengine==0.29.1
pymongo==4.10.1
PyJWT==2.8.0
dnspython==2.7.0
idna==3.10
python-decouple==3.8
razorpay==2.0.0
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0

# Production-specific packages
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.10